requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
redis==5.0.1
joblib==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.0.6
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2
optuna==3.5.0
ray==2.7.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
beautifulsoup4==4.12.2
lxml==5.1.0
tenacity==8.2.3
aiofiles==23.2.1
aioredis==2.0.1
httpx==0.26.0
apscheduler==3.10.4
aiohttp-retry==2.8.3
cachetools==5.3.2 